% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{routes}
\alias{routes}
\title{NYC Subway Routes}
\format{Data frame with columns
\describe{
\item{route_id}{Route ID}
\item{route_name}{Route name}
\item{route_color}{Route color hex code}
}}
\source{
\href{https://new.mta.info/developers/open-data}{Metropolitan Transport Authority}
}
\usage{
data(routes)
}
\description{
Data frame containing metadata on subway routes.
}
\keyword{datasets}
