% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stops}
\alias{stops}
\title{NYC Subway Stops}
\format{Data frame with columns
\describe{
\item{stop_id}{Stop ID}
\item{stop_name}{Stop name}
\item{stop_lat,stop_lon}{Stop latitude/longitude coordinates}
\item{stop_borough}{Borough containing stop}
}}
\source{
\href{https://new.mta.info/developers/open-data}{Metropolitan Transport Authority}
}
\usage{
data(stops)
}
\description{
Data frame containing metadata on subway stops.
}
\details{
Some stops emcompass multiple stations within a complex.
Such stops have coordinates equal to the centroid of stations at the complex.
}
\keyword{datasets}
