% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{travel_times}
\alias{travel_times}
\title{NYC Subway Travel Times}
\format{Data frame with columns
\describe{
\item{source}{Source stop ID}
\item{target}{Target stop ID}
\item{route_id}{Route ID}
\item{travel_time}{Minimum scheduled travel time between stops, in seconds}
}}
\source{
\href{https://new.mta.info/developers/open-data}{Metropolitan Transport Authority}
}
\usage{
data(travel_times)
}
\description{
Data frame containing minimum travel times between subway stops.
}
\details{
Each travel time represents the route-specific minimum difference between
the scheduled departure time from the source stop and the scheduled arrival
time at the target stop.
Travel times do not include delays between arrival and departure, nor
time taken to transfer between stations within a complex.
Therefore, summing travel times along a path provides a lower bound on the
total time required to traverse that path.
}
\keyword{datasets}
